created MentionList class
