+++ src/main/java/codeu/controller/ChatServlet.java
@@ -14,10 +14,7 @@
 
 package codeu.controller;
 
-import codeu.model.data.Conversation;
-import codeu.model.data.Mention;
-import codeu.model.data.Message;
-import codeu.model.data.User;
+import codeu.model.data.*;
 import codeu.model.store.basic.ConversationStore;
 import codeu.model.store.basic.MessageStore;
 import codeu.model.store.basic.UserStore;
@@ -161,88 +158,8 @@ public class ChatServlet extends HttpServlet {
     response.sendRedirect("/chat/" + conversationTitle);
 
     //Get mentions in an individual single message
-    List<Mention> mentionsList = getMentions(message);
+    MentionList mentionList = new MentionList(message, userStore);
 
   }
-  public List<Mention> getMentions(Message message){
-    List<Mention> mentions = new ArrayList<>();
-    String content = message.getContent();
 
-    Mention mention = null;
-    boolean inMention = false;
-    int startOfMention = -1;
-    int endOfMention = -1;
-
-    for (int i = 0; i < content.length(); i++) {
-
-      if(content.charAt(i)=='@'){
-        if(inMention){
-          inMention = false;
-          break;
-        }
-        else if(i==0 || content.charAt(i-1)==' '){
-          inMention = true;
-          startOfMention = i;
-        }
-
-        else if(content.charAt(i) == ' '){
-          if(inMention){
-            endOfMention = i;
-
-            mention = getMention(startOfMention, endOfMention, message);
-            if(!(mention == null)){
-              mentions.add(mention);
-            }
-            inMention = false;
-          }
-
-
-        }
-      }
-    }
-
-    return mentions;
-  }
-
-  public Mention getMention(int s, int e, Message m){
-    String content = m.getContent();
-    int start = s;
-    int end = e;
-    StringBuilder sb = new StringBuilder();
-
-    Mention mention = null;
-
-    for (int i = s; i < end; i++) {
-        sb.append(content.charAt(i));
-    }
-
-    String userMentioned = sb.toString();
-    UUID userMentionedID = searchForUser(userMentioned);
-
-    if(!(userMentionedID == null)){
-      mention = new Mention(
-              UUID.randomUUID(),
-              userMentionedID,
-              m.getAuthorId(),
-              start,
-              end,
-              m.getCreationTime(),
-              m);
-    }
-    return mention;
-  }
-
-  //Could be moved to UserStore
-  public UUID searchForUser(String username){
-    UUID userID = null;
-
-
-    for (User user : userStore.getAllUsers()) {
-      if(username.equalsIgnoreCase(user.getName())){
-        userID = user.getId();
-      }
-    }
-
-    return userID;
-  }
 }
