+++ src/main/java/codeu/model/data/Mention.java
@@ -55,4 +55,6 @@ public class Mention {
   }
 	/**Returns the message that the mention is in */
 	public Message getMessage(){return message;}
+
+
 }
